Switch to new Ask a Librarian UX
The web-based Ask a Librarian platform would enable tracking users
across every page on which the widget was embedded, so instead link
directly to the "Contact Us" page where it is embedded just once.

This also simplifies the CSS and JavaScript of the library home page
a little bit further.

Signed-off-by: Dan Scott <dan@coffeecode.net>